+++ src/components/Deposit/index.tsx
@@ -1,0 +1,2 @@
+export * from './DepositBlock'
+export * from './DepositDetailsCard'
